pyproject.toml
[project]
name = "agent-mcp-demos"
version = "0.1.0"
description = "Demo of agents using MCP tools"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Christoffer Björkskog", email = "christoffer.bjorkskog@novia.fi"}
]
license = {text = "MIT"}
keywords = ["mcp", "agent", "ai", "tools", "demo"]
dependencies = [
    "fastmcp>=2.12.5",
    "pydantic-ai>=1.3.0",
    "starlette>=0.48.0",
]
